Rename the db main file to index.
Index is more appropriate and allows teh imports to be neater

OHIE-376
